refactor: 使用 list_head 替代手动链表
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ daima/drivers/pet/pet_event.c
@@ -48,7 +48,7 @@ bool pet_build_chat_id(const char *chat_id, char *out, size_t out_size)
         chat_id += prefix_len;
     }
 
-    return snprintf(out, out_size, "%s%s", PET_CHAT_PREFIX, chat_id) < (int)out_size;
+    return snprintf(out, out_size, "%s%s", PET_CHAT_PREFIX, chat_id) < out_size;
 }
 
 bool pet_chat_id_to_ws_chat_id(const char *pet_chat_id, char *out, size_t out_size)
@@ -62,7 +62,7 @@ bool pet_chat_id_to_ws_chat_id(const char *pet_chat_id, char *out, size_t out_si
         return false;
     }
 
-    return strscpy(out, pet_chat_id + prefix_len, out_size) < (int)out_size;
+    return strscpy(out, pet_chat_id + prefix_len, out_size) < out_size;
 }
 
 size_t pet_append_channel_policy_prompt(char *prompt, size_t size, size_t offset)
